LayoutTests/platform/ios-sim/storage/indexeddb/versionchangerequest-activedomobject-expected.txt

CONSOLE MESSAGE: line 23: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
Ensure that IDBVersionChangeRequest objects are not GC'd if they have pending events

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;


self.dbname = 'versionchangerequest-activedomobject'

testDeleteDatabase():
indexedDB.deleteDatabase(self.dbname)
FAIL indexedDB.deleteDatabase(self.dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'deleteRequest.onerror = unexpectedErrorCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
